Pet Detail View added - able to add pet and cancel
Pet Detail View added - able to add pet and cancel with name, sex, dob and picture. Need to add ability to edit.

Also, it's a mystery why the outlets are not showing for delegate and datasource on the xCode UI. In fact, in the other project (PuppyVaccines), I only set these in xCode and did not need to add them programmatically.