Added the ability to Add and Cancel a new pet
At this point, I have a method for creating the data that first checks for iCloud and then loads 1) data from the cloud, 2) local data or 3) sample data. This is call for viewDidLoad and after each new record is created. Save action in Pets model saves locally and, if available, to iCloud. NOTE: this could cause inconsistency in the data and I need to work out my persistance strategy next.